// apps/redux-toolkit/src/state/queries/models/spotify.models.ts
export type SpotifyDatePrecision = 'year' | 'month' | 'day';

export type SpotifyAccessToken = {
  access_token: string;
  token_type: string;
  scope: string;
  expires_in: number;
  refresh_token: string;
};

type SpotifyExternalUrls = {
  spotify: string;
};

type SpotifyExternalIds = {
  isrc: string;
  ean: string;
  upc: string;
};

type SpotifyFollowers = {
  href?: string;
  total: number;
};

type SpotifyImage = {
  url: string;
  height?: number;
  width?: number;
};

type SpotifyRestrictions = {
  reason: string;
};

export type SpotifyCopyRights = {
  text: string;
  type: string;
};

type SpotifyExplicitContent = {
  filter_enabled: boolean;
  filter_locked: boolean;
};

type SpotifyCursors = {
  after: string;
  before: string;
};

export type SpotifyArtist = {
  external_urls: SpotifyExternalUrls;
  followers: SpotifyFollowers;
  genres: string[];
  href: string;
  id: string;
  images: SpotifyImage[];
  name: string;
  popularity: number;
  type: 'artist';
  uri: string;
};
export type SimplifiedSpotifyArtist = Pick<
  SpotifyArtist,
  'external_urls' | 'href' | 'id' | 'name' | 'type' | 'uri'
>;

export type SpotifyAlbum = {
  album_type: 'album' | 'single' | 'compilation';
  total_tracks: number;
  available_markets: string[];
  external_urls: SpotifyExternalUrls;
  href: string;
  id: string;
  images: SpotifyImage[];
  name: string;
  release_date: string;
  release_date_precision: SpotifyDatePrecision;
  restrictions: SpotifyRestrictions;
  type: 'album';
  uri: string;
  artists: SimplifiedSpotifyArtist[];
  tracks: SpotifyAlbumTracks;
  copyrights: SpotifyCopyRights[];
  external_ids: SpotifyExternalIds;
  genres: string[];
  label: string;
  popularity: number;
};
export type SimplifiedSpotifyAlbum = Omit<
  SpotifyAlbum,
  'tracks' | 'copyrights' | 'external_ids' | 'genres' | 'label' | 'popularity'
> & { album_group: 'album' | 'single' | 'compilation' | 'appears_on' };

type SpotifyAlbumTracks = {
  href: string;
  limit: number;
  next: string | null;
  offset: number;
  previous: string | null;
  total: number;
  items: SimplifiedSpotifyTrack[];
};

export type SpotifyTrack = {
  album: SpotifyAlbum;
  artists: SpotifyArtist[];
  available_markets: string[];
  disc_number: number;
  duration_ms: number;
  explicit: boolean;
  external_ids: SpotifyExternalIds;
  external_urls: SpotifyExternalUrls;
  href: string;
  id: string;
  is_playable: boolean;
  linked_from: unknown; //
  restrictions: SpotifyRestrictions;
  name: string;
  popularity: number;
  preview_url?: string;
  track_number: number;
  type: string;
  uri: string;
  is_local: boolean;
};
// ...
export type SpotifyDevice = {
  id: string | null;
  is_active: boolean;
  is_private_session: boolean;
  is_restricted: boolean;
  name: string;
  type: string;
  volume_percent: number;
  supports_volume: boolean;
};

type SpotifyContext = {
  type: 'artist' | 'playlist' | 'album' | 'show';
  href: string;
  external_urls: SpotifyExternalUrls;
  uri: string;
};

type SpotifyResumePoint = {
  fully_played: boolean;
  resume_position_ms: number;
};

type SpotifyShow = {
  available_markets: string[];
  copyrights: SpotifyCopyRights;
  description: string;
  html_description: string;
  explicit: boolean;
  external_urls: SpotifyExternalUrls;
  href: string;
  id: string;
  images: SpotifyImage[];
  is_externally_hosted: boolean;
  languages: string[];
  media_type: string;
  name: string;
  publisher: string;
  type: 'show';
  uri: string;
  total_episodes: number;
};

type SpotifyEpisode = {
  audio_preview_url: string | null;
  description: string;
  html_description: string;
  duration_ms: number;
  explicit: boolean;
  external_urls: SpotifyExternalUrls;
  href: string;
  id: string;
  images: SpotifyImage[];
  is_externally_hosted: boolean;
  is_playable: boolean;
  languages: string[];
  name: string;
  release_date: string;
  release_date_precision: SpotifyDatePrecision;
  resume_point: SpotifyResumePoint;
  type: 'episode';
  uri: string;
  restrictions: SpotifyRestrictions;
  show: SpotifyShow;
};

type SpotifyPlaybackActions = {
  interrupting_playback: boolean;
  pausing: boolean;
  resuming: boolean;
  seeking: boolean;
  skipping_next: boolean;
  skipping_prev: boolean;
  toggling_repeat_context: boolean;
  toggling_shuffle: boolean;
  toggling_repeat_track: boolean;
  transferring_playback: boolean;
};

export type SpotifyPlaybackState = {
  device: SpotifyDevice;
  repeat_state: 'off' | 'track' | 'context';
  shuffle_state: boolean;
  context: SpotifyContext;
  timestamp: number;
  progress_ms: number;
  is_playing: boolean;
  item: SpotifyTrack | SpotifyEpisode;
  currently_playing_type: 'track' | 'episode' | 'unknown';
  actions: SpotifyPlaybackActions;
};